move get_type and prototype declarations
