+++ src/code_generator/type.c
@@ -1,0 +1,14 @@
+#include "code_generator.h"
+#include "stdstring.h"
+#include "ast/is_method.h"
+#include "ast/get_method.h"
+#include "ast/ast_impl.h"
+
+LLVMTypeRef get_type(AstRef ast) {
+  switch (ast->tag) {
+    case AST_DECLARATION_SPECIFIER_LIST:
+      return LLVMInt32Type();
+    default:
+      return NULL;
+  }
+}
